Move loading of PostgreSQLProxy to its own Railtie
