+++ lib/active_record_proxy_adapters/connection_handling/postgresql_proxy.rb
