+++ lib/active_record_proxy_adapters/railtie.rb
@@ -1,18 +1,18 @@
 # frozen_string_literal: true
 
 require "active_support"
+require "active_record_proxy_adapters/railties/postgresql"
 
 module ActiveRecordProxyAdapters
   # Hooks into rails boot process to extend ActiveRecord with the proxy adapter.
   class Railtie < Rails::Railtie
-    require "active_record_proxy_adapters/connection_handling"
+    require "active_record_proxy_adapters/connection_handling/mysql2"
+    require "active_record_proxy_adapters/connection_handling/trilogy"
+    require "active_record_proxy_adapters/connection_handling/sqlite3"
     require "active_record_proxy_adapters/middleware"
 
     config.to_prepare do
       Rails.autoloaders.each do |autoloader|
-        autoloader.inflector.inflect(
-          "postgresql_proxy_adapter" => "PostgreSQLProxyAdapter"
-        )
         autoloader.inflector.inflect(
           "sqlite3_proxy_adapter" => "SQLite3ProxyAdapter"
         )
